api/requirements.txt
djangorestframework

markdown

django-filter

django

